refactor(bulkupload): Move restore/load entities into own file
